Programming with UNumber
Programming with UNumber; Implementing Square Root with UNumber*
(*Unumber- It is a user defined library)
The goal of is to leverage the experiences gathered so far in order to begin the process of designing an enhancement to the UNumber library to support square root.

The outcome is a concrete list of things that must be done in order to have square root be a solid addition to the library.